+++ CHANGELOG.md
@@ -12,6 +12,7 @@ ChangeLog
 - BUGFIX: fix resource handling by tracking pending promises in "a2a" nodes
 - BUGFIX: fix race condition between chunkQueue nulling and stream destruction in MQTT and VBAN nodes
 - BUGFIX: fix start timestamp calculation in "a2t-google" node
+- BUGFIX: fix error message in case a positional parameter does not match
 - UPDATE: upgrade NPM dependencies and patches
 - CLEANUP: add missing type definitions for Kitten-TTS
 - CLEANUP: update year in all copyright messages